requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-community
langchain-openai
sentence-transformers
faiss-cpu
openai